Fix various compilation warnings for gcc-4.2.
Approved by:	re (bruce)


Former-commit-id: 5b27eab7fcb14e23315a746c02059e069c84c967

+++ sys/dev/sym/sym_hipd.c
@@ -627,7 +627,7 @@ parameter_list|,
 name|member
 parameter_list|)
 define|\
-value|((type *)((char *)(ptr)-(unsigned int)(&((type *)0)->member)))
+value|((type *)((char *)(ptr)-(size_t)(&((type *)0)->member)))
 end_define
 
 begin_define
